Refactor JVM apps to use shared components
